Fix compile warning

+++ XDMA/linux-kernel/xdma/xdma_netdev.c
@@ -107,7 +107,7 @@ netdev_tx_t xdma_netdev_start_xmit(struct sk_buff *skb,
         /* Check desc count */
         netif_stop_queue(ndev);
         xdma_debug("xdma_netdev_start_xmit(skb->len : %d)\n", skb->len);
-        skb->len = max(ETH_ZLEN, skb->len);
+        skb->len = max((unsigned int)ETH_ZLEN, skb->len);
         if (skb_padto(skb, skb->len)) {
                 pr_err("skb_padto failed\n");
                 dev_kfree_skb(skb);
